[Unification] Move all encoders/transformers to models folders
[ghstack-poisoned]
