+++ test/models/clip/test_clip.py
@@ -7,9 +7,9 @@
 import pytest
 import torch
 from test.test_utils import assert_expected, set_rng_seed
-from torchmultimodal.models.clip import CLIP
-from torchmultimodal.modules.encoders.clip_resnet_encoder import ResNetForCLIP
-from torchmultimodal.modules.encoders.clip_text_encoder import CLIPTextEncoder
+from torchmultimodal.models.clip.model import CLIP
+from torchmultimodal.models.clip.resnet_encoder import ResNetForCLIP
+from torchmultimodal.models.clip.text_encoder import CLIPTextEncoder
 from torchvision.models.vision_transformer import VisionTransformer
