+++ torchmultimodal/models/albef/image_encoder.py
