+++ torchmultimodal/models/clip/model.py
@@ -11,8 +11,8 @@ import torch
 import torch.nn.functional as F
 from torch import nn
 
-from torchmultimodal.modules.encoders.clip_resnet_encoder import ResNetForCLIP
-from torchmultimodal.modules.encoders.clip_text_encoder import CLIPTextEncoder
+from torchmultimodal.models.clip.resnet_encoder import ResNetForCLIP
+from torchmultimodal.models.clip.text_encoder import CLIPTextEncoder
 from torchvision.models.resnet import Bottleneck, ResNet
 from torchvision.models.vision_transformer import VisionTransformer
