+++ torchmultimodal/models/clip/resnet_encoder.py
