+++ torchmultimodal/models/clip/text_encoder.py
